feat(booking): S3-READ-DB

+++ contracts/src/main/java/com/team7/eventticketing/contracts/dto/BookingDTO.java
@@ -1,11 +1,10 @@
 package com.team7.eventticketing.contracts.dto;
 
-
 import com.team7.eventticketing.contracts.enums.BookingStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
-
 
 public record BookingDTO(
         Long id,
@@ -13,7 +12,7 @@ public record BookingDTO(
         Long eventId,
         String contactEmail,
         BookingStatus status,
-        Double totalAmount,
+        BigDecimal totalAmount,
         Map<String, Object> metadata,
         LocalDateTime bookingDate,
         LocalDateTime confirmedAt,
