+++ contracts/src/main/java/com/team7/eventticketing/contracts/dto/EventTicketSummaryDTO.java
@@ -1,3 +1,3 @@
 package com.team7.eventticketing.contracts.dto;
 
-public record EventTicketSummaryDTO(int totalTicketsSold, int usedCount) {}
+public record EventTicketSummaryDTO(long totalTicketsSold, int usedCount) {}
